requirements.txt
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==1.11.6
django-cors-headers==2.1.0
django-crispy-forms==1.7.0
django-filter==1.1.0
django-formtools==2.1
django-guardian==1.4.9
django-redis==4.8.0
django-reversion==2.0.12
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
drf-extensions==0.3.1
future==0.16.0
httplib2==0.10.3
idna==2.6
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.10
MarkupSafe==1.0
mysqlclient==1.3.12
oauthlib==2.0.6
olefile==0.44
Pillow==4.3.0
pycryptodome==3.4.7
PyJWT==1.5.3
PyMySQL==0.7.11
python3-openid==3.1.0
pytz==2017.3
raven==6.4.0
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
uritemplate==3.0.0
urllib3==1.22
XlsxWriter==1.0.2
xlwt==1.3.0
